PR Review Changes issues resolved.
